+++ R/get_all_by_isotope.R
@@ -2,13 +2,21 @@
 #'
 #' @param qsip_data_object (*qsip_data or qsip_source_data*) A qsip object with source data
 #' @param isotopes (*string(s)*) Isotopes used to pull source_mat_ids. Can be a standard isotope name (e.g. `12C`) or special terms `labeled` or `unlabeled`
-#' @param silent (*boolean*) If `TRUE`, suppresses messages about missing isotope hits and doesn't fail
+#' @param quiet (*boolean*) If `TRUE`, suppresses messages about missing isotope hits and doesn't fail
 #'
 #' @returns A vector of source_mat_ids. It may also print some messages.
 #'
 #' @export
 
-get_all_by_isotope <- function(qsip_data_object, isotopes, silent = FALSE) {
+get_all_by_isotope <- function(qsip_data_object,
+                               isotopes,
+                               quiet = FALSE,
+                               silent = lifecycle::deprecated()) {
+
+  if (lifecycle::is_present(silent)) {
+    lifecycle::deprecate_warn("0.18.4.9000", "get_all_by_isotope(silent)")
+  }
+
   if (inherits(qsip_data_object, qsip_data)) {
     source_data <- qsip_data_object@source_data@data
   } else if (inherits(qsip_data_object, qsip_source_data)) {
@@ -36,7 +44,7 @@ get_all_by_isotope <- function(qsip_data_object, isotopes, silent = FALSE) {
     dplyr::select(source_mat_id, isotope)
 
   # error if no source_mat_ids are found that match the criteria
-  if (isFALSE(silent)) {
+  if (isFALSE(quiet)) {
 
     if (nrow(source_mat_ids) == 0) {
       i <- paste(isotopes, collapse = ", ")